Add new azure flag to obtain groups from the oidc ticket
It is possible in azure to configure an optional claim such as groups.
In this case there is no need to make a request to Microsoft Graph api
to obtain groups. Instead azure will send the configured groups as a
part of oidc ticket directly. This is usefull when it is not possible to
obtain GroupMember.Read.All permissions for the applictaion.

Here we provide a new flag `azure-groups-in-ticket`, which does exactly
this. When the flag is set the request to the Graph api is not done and
instead the groups are taken from the session directly.